fix: solve bug04

+++ src/main/java/com/amigoscode/bughunt/easy/bug04/Percentage.java
@@ -2,7 +2,7 @@ package com.amigoscode.bughunt.easy.bug04;
 
 public class Percentage {
 
-    public double percentage(int part, int total) {
+    public double percentage(double part, int total) {
         return (part / total) * 100;
     }
 }
